refactor: extract ErrorBoundary component and add source docs
